Rename benchmark.py to benchmarks.py.
