+++ tests/templates/conftest.py
@@ -3,7 +3,7 @@ import pathlib
 import pytest
 from jinja2 import Environment, FileSystemLoader
 
-from coffee.benchmark import (
+from coffee.benchmarks import (
     GeneralPurposeAiChatBenchmark,
     BenchmarkScore,
 )
